# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JLPT Mobile API",
        "description": "REST API for the JLPT mobile app. All authenticated endpoints expect a Sanctum bearer token issued by /auth/login or /auth/register.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://jlpt.senda.fit",
            "description": "Production"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local"
        }
    ],
    "paths": {
        "/api/v1/auth/email/verify-send": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Resend the email verification link to the authenticated user",
                "operationId": "5d032c70d8c86f3619b9d0b7f9d6dcf4",
                "responses": {
                    "200": {
                        "description": "Verification email sent"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/auth/email/verify/{id}/{hash}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Confirm email via the signed link sent over email",
                "operationId": "d3488e1ddfc474c63037c03f7700c4e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "hash",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email verified"
                    },
                    "403": {
                        "description": "Invalid verification link"
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login with email and password",
                "operationId": "02f6ee354f76eb020df20e4d9b68bbb0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "device_name"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "device_name": {
                                        "type": "string",
                                        "example": "iphone-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/auth/forgot-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Email a password reset link",
                "operationId": "0a7dba788c42aa3703b2233c4340cbd2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reset link sent"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/auth/reset-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Reset password using the emailed token",
                "operationId": "8287fc6f8bbb43dd0b1932319b713607",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset"
                    },
                    "422": {
                        "description": "Invalid token or password"
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new app user with email and password",
                "operationId": "b3e30f5d550777b9cf8cddeceaf2c986",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "device_name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Sulaiman MD"
                                    },
                                    "display_name": {
                                        "type": "string",
                                        "example": "smensulaiman",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "device_name": {
                                        "type": "string",
                                        "example": "iphone-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AppUser": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    },
                    "name": {
                        "type": "string",
                        "example": "Sulaiman MD"
                    },
                    "username": {
                        "type": "string",
                        "example": "smensulaiman",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "email_verified": {
                        "type": "boolean"
                    },
                    "avatar": {
                        "type": "string",
                        "nullable": true
                    },
                    "login_provider": {
                        "type": "string",
                        "enum": [
                            "email",
                            "google",
                            "apple",
                            "facebook",
                            "guest"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "bio": {
                        "type": "string",
                        "nullable": true
                    },
                    "gender": {
                        "type": "string",
                        "nullable": true,
                        "enum": [
                            "male",
                            "female",
                            "other",
                            "prefer_not_to_say"
                        ]
                    },
                    "birth_date": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "native_language_code": {
                        "type": "string"
                    },
                    "learning_language_code": {
                        "type": "string"
                    },
                    "proficiency_level": {
                        "type": "string",
                        "enum": [
                            "beginner",
                            "elementary",
                            "intermediate",
                            "upper_intermediate",
                            "advanced",
                            "native_like"
                        ]
                    },
                    "learning_goal": {
                        "type": "string",
                        "enum": [
                            "travel",
                            "school",
                            "business",
                            "anime",
                            "conversation",
                            "exam",
                            "hobby",
                            "other"
                        ]
                    },
                    "daily_goal_minutes": {
                        "type": "integer"
                    },
                    "timezone": {
                        "type": "string",
                        "nullable": true
                    },
                    "reminder_time": {
                        "type": "string",
                        "nullable": true
                    },
                    "notifications_enabled": {
                        "type": "boolean"
                    },
                    "coins": {
                        "type": "integer"
                    },
                    "xp_points": {
                        "type": "integer"
                    },
                    "current_level": {
                        "type": "integer"
                    },
                    "current_streak": {
                        "type": "integer"
                    },
                    "longest_streak": {
                        "type": "integer"
                    },
                    "learned_words": {
                        "type": "integer"
                    },
                    "completed_lessons": {
                        "type": "integer"
                    },
                    "total_study_minutes": {
                        "type": "integer"
                    },
                    "last_streak_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "last_seen_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "last_lesson_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "favorites_count": {
                        "type": "integer"
                    },
                    "reads_count": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "AuthResponse": {
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/AppUser"
                    },
                    "token": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Sanctum token",
                "scheme": "bearer"
            }
        }
    },
...
}